shrine-api/requirements.txt
basho-erlastic==2.1.1
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
python3-protobuf==2.5.0
PyYAML==3.13
riak==2.7.0
six==1.12.0
uWSGI==2.0.18
Werkzeug==0.14.1
